==> analyses/independent-samples/03-qc-independent-samples.Rmd <==
---
title: "QC independent samples"
output: 
  html_notebook:
    toc: true
    toc_float: true
---

## Load libraries

```{r load_libraries}
suppressPackageStartupMessages({
  library(tidyverse)
  library(DT)
})
```

## Output files

```{r wgs_only}
wgs_only_files <- c("independent-specimens.wgs.primary.eachcohort.tsv",
                    "independent-specimens.wgs.primary.tsv",
                    "independent-specimens.wgs.relapse.eachcohort.tsv",
                    "independent-specimens.wgs.relapse.tsv",
                    "independent-specimens.wgs.primary-plus.eachcohort.tsv",
                    "independent-specimens.wgs.primary-plus.tsv")
```

```{r wgs_preferred}
wgs_preferred_files <- c("independent-specimens.wgswxspanel.primary.eachcohort.prefer.wgs.tsv",
                         "independent-specimens.wgswxspanel.primary.prefer.wgs.tsv",
                         "independent-specimens.wgswxspanel.relapse.eachcohort.prefer.wgs.tsv",
                         "independent-specimens.wgswxspanel.relapse.prefer.wgs.tsv",
                         "independent-specimens.wgswxspanel.primary-plus.eachcohort.prefer.wgs.tsv",
                         "independent-specimens.wgswxspanel.primary-plus.prefer.wgs.tsv")
```

```{r wxs_preferred}
wxs_preferred_files <- c("independent-specimens.wgswxspanel.primary.eachcohort.prefer.wxs.tsv",
                         "independent-specimens.wgswxspanel.primary.prefer.wxs.tsv",
                         "independent-specimens.wgswxspanel.relapse.eachcohort.prefer.wxs.tsv",
                         "independent-specimens.wgswxspanel.relapse.prefer.wxs.tsv",
                         "independent-specimens.wgswxspanel.primary-plus.eachcohort.prefer.wxs.tsv",
                         "independent-specimens.wgswxspanel.primary-plus.prefer.wxs.tsv")
```

```{r rnaseq}
rnaseq_files <- c("independent-specimens.rnaseqpanel.primary.eachcohort.tsv",
                         "independent-specimens.rnaseqpanel.primary.tsv",
                         "independent-specimens.rnaseqpanel.relapse.eachcohort.tsv",
                         "independent-specimens.rnaseqpanel.relapse.tsv",
                         "independent-specimens.rnaseqpanel.primary-plus.eachcohort.tsv",
                         "independent-specimens.rnaseqpanel.primary-plus.tsv")
```

```{r methyl}
methyl_files <- c("independent-specimens.methyl.primary.tsv",
                         "independent-specimens.methyl.relapse.tsv",
                         "independent-specimens.methyl.primary-plus.tsv")
```

## Functions

```{r set_functions}
# number of total rows + number of total Kids_First_Participant_ID
count_num <- function(x){
  df <- read_tsv(x, col_types = cols())
  data.frame(file = x, n_rows = nrow(df), unique_KF_id = length(unique(df$Kids_First_Participant_ID)))
}
```

### WGS File summary

```{r wgs_file_summary}
DT::datatable(map_df(wgs_only_files, .f = function(x) 
  count_num(file.path('results', x))))
```

### WGS-preferred File summary

```{r wgs_pref_file_summary}
DT::datatable(map_df(wgs_preferred_files, .f = function(x) 
  count_num(file.path('results', x))))
```

### WXS-preferred File summary

```{r wxs_pref_file_summary}
DT::datatable(map_df(wxs_preferred_files, .f = function(x) 
  count_num(file.path('results', x))))
```

### RNA-Seq File summary

```{r rnaseq_file_summary}
DT::datatable(map_df(rnaseq_files, .f = function(x) 
  count_num(file.path('results', x))))
```

### Methylation File summary

```{r methyl_file_summary}
DT::datatable(map_df(methyl_files, .f = function(x) 
  count_num(file.path('results', x))))
```
